requirements.txt
fastapi
uvicorn
pydantic
google-genai
certifi
webrtcvad
python-multipart
websockets
httpx
numpy
piper-tts
python-dotenv
